backend/requirements.txt
asgiref==3.11.1
asttokens==3.0.1
decorator==5.2.1
dj-database-url==3.1.2
Django==6.0.3
django-admin-interface==0.32.0
django-colorfield==0.14.0
django-flat-responsive==2.0
django-jet-reboot==1.3.10
django-cors-headers==4.3.1
django-unfold==0.85.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
dotenv==0.9.9
executing==2.2.1
ipython==9.11.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
matplotlib-inline==0.2.1
parso==0.8.6
pexpect==4.9.0
pillow==12.1.1
prompt_toolkit==3.0.52
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
python-dotenv==1.2.2
python-slugify==8.0.4
sqlparse==0.5.5
stack-data==0.6.3
text-unidecode==1.3
traitlets==5.14.3
wcwidth==0.6.0
redis==5.0.3
channels-redis==4.2.0
channels==4.2.2
daphne==4.1.2
